Rename api to odm2_rest_api with a goal to create reusable app
