+++ .travis.yml
@@ -63,9 +63,9 @@ before_install:
 
 script:
   - if [[ $TEST_TARGET == 'default' ]]; then
-      python $TRAVIS_BUILD_DIR/manage.py test api;
+      python $TRAVIS_BUILD_DIR/manage.py test odm2_rest_api;
     fi
   - if [[ $TEST_TARGET == 'coding_standards' ]]; then
-      flake8 --max-line-length=105 api;
+      flake8 --max-line-length=105 odm2_rest_api;
       flake8 --max-line-length=105 odm2rest;
     fi
