+++ odm2_rest_api/apps.py
@@ -5,4 +5,4 @@ from django.apps import AppConfig
 
 
 class ApiConfig(AppConfig):
-    name = 'api'
+    name = 'odm2_rest_api'
