+++ odm2_rest_api/static/odm2_rest_api/js/swagger-ui.js
