+++ odm2_rest_api/static/odm2_rest_api/js/swagger.json
